pyproject.toml
[tool.poetry]
name = "aimarkerspace-middleterm"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11.7,<3.12"
llama-index = "^0.10.18"
openai = "^1.13.3"
wandb = "^0.16.4"
llama-index-callbacks-wandb = "^0.1.2"
faiss-cpu = "^1.8.0"
pymupdf = "^1.23.26"
tiktoken = "0.4.0"
sentence-transformers = "2.2.2"
pydantic = "1.10.11"
llama-index-vector-stores-faiss = "^0.1.2"
chainlit = "0.7.700"
python-dotenv = "1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
